requirements.txt
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
greenlet==1.1.3
idna==3.4
numpy==1.23.3
packaging==21.3
pandas==1.5.0
prettytable==3.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
SQLAlchemy==1.4.41
urllib3==1.26.12
wcwidth==0.2.5
wrapt==1.14.1
